se cambio el quantitypercent en int

+++ app/src/main/java/com/win/gestionderiesgos/domain/repository/registerFuncions/RegisterFuncionRepository.kt
@@ -28,7 +28,7 @@ class RegisterFuncionRepository {
                         val idUser =ds.child("idUser").value.toString()
                         val dateCreated =ds.child("dateCreated").value.toString()
                         val QuantityPercent=ds.child("QuantityPercent").value.toString()
-                        val listas =Funcions(name,idUser, dateCreated,QuantityPercent)
+                        val listas =Funcions(name,idUser, dateCreated,QuantityPercent.toInt())
                         listFuncions.add(listas)
                     }
                     mutableLiveData.value =listFuncions
